refactor: extract download logic in `spack.util.downloader`
Consolidates download-related functionality into a dedicated
module to improve maintainability and modularity.

Updates affected components and tests to use the new
implementation.
