+++ lib/spack/spack/test/fetch_strategy.py
@@ -7,6 +7,7 @@ from io import StringIO
 import pytest
 
 from spack import fetch_strategy
+from spack.util import downloader
 
 
 def test_fetchstrategy_bad_url_scheme():
@@ -32,7 +33,7 @@ def test_fetchstrategy_bad_url_scheme():
     ],
 )
 def test_format_bytes(expected, total_bytes):
-    assert fetch_strategy._format_bytes(total_bytes) == expected
+    assert downloader._format_bytes(total_bytes) == expected
 
 
 @pytest.mark.parametrize(
@@ -53,13 +54,13 @@ def test_format_bytes(expected, total_bytes):
     ],
 )
 def test_format_speed(expected, total_bytes, elapsed):
-    assert fetch_strategy._format_speed(total_bytes, elapsed) == expected
+    assert downloader._format_speed(total_bytes, elapsed) == expected
 
 
 def test_fetch_progress_unknown_size():
     # time stamps in seconds, with 0.1s delta except 1.5 -> 1.55.
     time_stamps = iter([1.0, 1.5, 1.55, 2.0, 3.0, 5.0, 5.5, 5.5])
-    progress = fetch_strategy.FetchProgress(total_bytes=None, get_time=lambda: next(time_stamps))
+    progress = downloader.FetchProgress(total_bytes=None, get_time=lambda: next(time_stamps))
     assert progress.start_time == 1.0
     out = StringIO()
 
@@ -91,7 +92,7 @@ def test_fetch_progress_unknown_size():
 
 def test_fetch_progress_known_size():
     time_stamps = iter([1.0, 1.5, 3.0, 4.0, 4.0])
-    progress = fetch_strategy.FetchProgress(total_bytes=6000, get_time=lambda: next(time_stamps))
+    progress = downloader.FetchProgress(total_bytes=6000, get_time=lambda: next(time_stamps))
     out = StringIO()
     progress.advance(1000, out)  # time 1.5
     progress.advance(2000, out)  # time 3.0
@@ -114,7 +115,7 @@ def test_fetch_progress_disabled():
     def get_time():
         raise RuntimeError("Should not be called")
 
-    progress = fetch_strategy.FetchProgress(enabled=False, get_time=get_time)
+    progress = downloader.FetchProgress(enabled=False, get_time=get_time)
     out = StringIO()
     progress.advance(1000, out)
     progress.advance(2000, out)
@@ -135,7 +136,7 @@ def test_fetch_progress_disabled():
 )
 def test_fetch_progress_from_headers(header, value, total_bytes):
     time_stamps = iter([1.0, 1.5, 3.0, 4.0, 4.0])
-    progress = fetch_strategy.FetchProgress.from_headers(
+    progress = downloader.FetchProgress.from_headers(
         {header: value}, get_time=lambda: next(time_stamps), enabled=True
     )
     assert progress.total_bytes == total_bytes
@@ -144,7 +145,7 @@ def test_fetch_progress_from_headers(header, value, total_bytes):
 
 
 def test_fetch_progress_from_headers_disabled():
-    progress = fetch_strategy.FetchProgress.from_headers(
+    progress = downloader.FetchProgress.from_headers(
         {"Content-Length": "1234"}, get_time=lambda: 1.0, enabled=False
     )
     assert not progress.enabled
